requirements.txt
mysql-connector-python
Faker
kafka-python
python-dotenv
loguru
pycountry
phonenumbers
postal 